+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ## Added
 - API method get_filesystem_utilization
 - Automatically-generated uploaded parameter to gcode files
+- Option to configure lcd_gpio_pin for filament detection
 
 ## [0.16.1] - 2016-04-09
 ## Fixed